requirements.txt
absl-py==0.7.1
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cityscapesScripts==1.4.0
cloudpickle==1.2.1
conda==4.5.11
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.20
dask==1.2.2
decorator==4.4.0
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
future==0.18.2
futures==3.2.0
gast==0.3.2
grpcio==1.21.1
h5py==2.8.0
idna==2.7
imageio==2.5.0
imgaug==0.2.9
imutils==0.5.3
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
jedi==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.3
MedPy==0.4.0
mock==3.0.5
networkx==2.2
numpy==1.15.1
nvidia-ml-py==375.53.1
opencv-python==3.4.2.17
pandas==0.23.4
parso==0.5.0
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==1.0.16
protobuf==3.8.0
ptyprocess==0.6.0
pycocotools==12.0
pycosat==0.6.3
pycparser==2.18
pydicom==1.0.2
Pygments==2.4.2
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.19.1
ruamel-yaml==0.15.46
scandir==1.10.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
SimpleITK==1.2.0
six==1.13.0
split-folders==0.2.2
stomp.py==4.1.22
subprocess32==3.5.4
tensorboard==1.11.0
tensorboardX==2.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
toolz==0.9.0
torch==1.1.0
torchvision==0.2.0
tqdm==4.43.0
traitlets==4.3.2
urllib3==1.23
utils==0.9.0
wcwidth==0.1.7
Werkzeug==0.15.4
